[MOBI-21-landing-page-top-desktop-publishing] Feat: custom keyframe을 className으로 적용합니다.
