--- apps/academy/src/feature/imageSlider/constants/index.ts
@@ -1,1 +1,0 @@
-export { MOBI_DESCRIPTION_IMAGES } from './MobiDescription'
