--- apps/academy/src/feature/imageSlider/lib/index.ts
@@ -1,1 +1,0 @@
-export { useImageSlider } from './use-image-slider'
