--- apps/academy/src/feature/imageSlider/ui/index.ts
@@ -1,1 +1,0 @@
-export { MobiDescriptionImagesSlider } from './MobiDescriptionImagesSlider'
