+++ apps/academy/src/widget/landing/constants/MobiDescription.ts
@@ -40,3 +40,17 @@ export const MOBI_DESCRIPTION = [
     ]
   }
 ]
+
+export type MobIDescriptionImageType = {
+  id: number
+  src: string
+}
+
+export const MOBI_DESCRIPTION_IMAGES: MobIDescriptionImageType[] = [
+  { id: 1, src: '/webp/landing/landing-top-image-1.webp' },
+  { id: 2, src: '/webp/landing/landing-top-image-2.webp' },
+  { id: 3, src: '/webp/landing/landing-top-image-3.webp' },
+  { id: 4, src: '/webp/landing/landing-top-image-4.webp' },
+  { id: 5, src: '/webp/landing/landing-top-image-5.webp' },
+  { id: 6, src: '/webp/landing/landing-top-image-6.webp' }
+]
