+++ apps/academy/src/widget/landing/constants/index.ts
@@ -1,1 +1,1 @@
-export { MOBI_DESCRIPTION } from './MobiDescription'
+export { MOBI_DESCRIPTION, MOBI_DESCRIPTION_IMAGES } from './MobiDescription'
